Perform initial cleanup of survey frontend
