--- app/uk/gov/hmrc/paymentssurvey/controllers/SurveyController.scala
@@ -1,5 +1,0 @@
-package uk.gov.hmrc.paymentssurvey.controllers
-
-class SurveyController {
-
-}
