+++ conf/app.routes
@@ -1,3 +1,2 @@
 # microservice specific routes
 
-GET        /hello-world             uk.gov.hmrc.paymentssurvey.controllers.HelloWorld.helloWorld
