+++ conf/prod.routes
@@ -1,5 +1,5 @@
 # Add all the application routes to the app.routes file
-->         /payments-survey              app.Routes
+->         /payments-survey           app.Routes
 ->         /                          health.Routes
 ->     	   /template                  template.Routes
