--- it/uk/gov/hmrc/paymentssurvey/ExampleIntegrationTest.scala
@@ -1,24 +1,0 @@
-package uk.gov.hmrc.paymentssurvey
-
-import org.scalatest.{Matchers, WordSpec}
-import play.api.libs.ws.WSClient
-import uk.gov.hmrc.integration.ServiceSpec
-
-class ExampleIntegrationTest extends WordSpec with Matchers with ServiceSpec  {
-
-  def externalServices: Seq[String] = Seq("datastream", "auth")
-
-  override def additionalConfig: Map[String, _] = Map("auditing.consumer.baseUri.port" -> externalServicePorts("datastream"))
-
-
-  "This integration test" should {
-    "start services via smserver" in {
-
-      val wsClient = app.injector.instanceOf[WSClient]
-
-      val response = wsClient.url(resource("/payments-survey/hello-world")).get.futureValue
-      response.status shouldBe 200
-
-    }
-  }
-}
