+++ project/plugins.sbt
@@ -5,13 +5,20 @@ resolvers += "HMRC Releases" at "https://dl.bintray.com/hmrc/releases"
 
 resolvers += Resolver.typesafeRepo("releases")
 
+addSbtPlugin("com.typesafe.play" % "sbt-plugin" % "2.6.25")
 
-addSbtPlugin("uk.gov.hmrc" % "sbt-auto-build" % "1.13.0")
+addSbtPlugin("org.scalariform" % "sbt-scalariform" % "1.8.2")
 
-addSbtPlugin("uk.gov.hmrc" % "sbt-git-versioning" % "1.15.0")
+addSbtPlugin("org.scoverage" % "sbt-scoverage" % "1.6.1")
 
-addSbtPlugin("uk.gov.hmrc" % "sbt-artifactory" % "0.13.0")
+addSbtPlugin("org.wartremover" % "sbt-wartremover" % "2.3.1")
 
-addSbtPlugin("uk.gov.hmrc" % "sbt-distributables" % "1.1.0")
+addSbtPlugin("uk.gov.hmrc" % "sbt-artifactory" % "1.0.0")
 
-addSbtPlugin("com.typesafe.play" % "sbt-plugin" % "2.5.12")
+addSbtPlugin("uk.gov.hmrc" % "sbt-auto-build" % "2.5.0")
+
+addSbtPlugin("uk.gov.hmrc" % "sbt-bobby" % "2.4.0")
+
+addSbtPlugin("uk.gov.hmrc" % "sbt-distributables" % "2.0.0")
+
+addSbtPlugin("uk.gov.hmrc" % "sbt-git-versioning" % "2.1.0")
